+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 1.0.0 Unreleased
+ * [MODORDERS-79](https://issues.folio.org/browse/MODORDERS-79) - Implemented get order line logic
  * [MODORDERS-75](https://issues.folio.org/browse/MODORDERS-75) - Define order line endpoints
  * [MODORDERS-62](https://issues.folio.org/browse/MODORDERS-62) - Migrate to RAML1.0 and RMB 23
  * [MODORDERS-51](https://issues.folio.org/browse/MODORDERS-51)